move some methods from watcher to AvaFiles
